Add emergency break to position controller
If position leaves the defined range, an error is risen, which will
cause a "disable system" command to be sent.
Note that this is only a hacky solution, as it assumes that the leg is
initialized in fully streched position (i.e. zero lies in the middle of
the motor ranges).