Enable calibration and initialization of zero position
Add a new script `calibrate.py` that is used for simple calibration of
the stop position to zero position offset.  The result is stored in a
file which can then be used by the new function `init_position_offset()`
to easily initialize the leg position at startup in every script.
Subtracting the offest is then done inside MotorData, so it is
completely transparent to the user.