Make time portable
time.clock() behaves differently on windows and linux. While on Windows
clock() is preferred over time() due to its higher accuracy, on Linux
clock() does return CPU time instead of wall time.  Therefore use
clock() when on windows and time() otherwise (this is how it is done in
timeit as well).